+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * newblock.sh block importer: Add accordion
 * newblock.sh block importer: Add carousel
 * New feature: Activates the block ACF fields automatically (no extra step with "Sync available" > Import)
+* New feature: You can now add new empty blocks!
 
 ### 1.0.3: 2021-02-08